Add 'engagement' and 'philosophy' articles
The "scripture-engagement-scale-placeholder.png" image was a screen-shot
from Figma and may need to be replaced with a better version.

I haven't added the "Resource Recommendations" article partly because
there is a note to reword some of the text. The styles are a little
more complex than I thought possible with simple Markdown, so I'll
explore that in more detail later.