Change links to dark salmon color
Addresses comment (and all related comments):

David Hammaker Noah This seems nitpicky, but I don't want to risk getting flagged for being non-compliant. Could we update this salmon color to match the darker salmon on the design (frame on the left)
Actually, all in-line text links on the light background should be this darker shade of salmon to follow compliance. Thank you!